requirements.txt
discord.py==2.3.0
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0